Move GPUTypeChoices, schema and function to components.schemas
