+++ app/grandchallenge/algorithms/models.py
@@ -40,7 +40,7 @@ from grandchallenge.components.models import (  # noqa: F401
     ImportStatusChoices,
     Tarball,
 )
-from grandchallenge.components.utils import GPUTypeChoices
+from grandchallenge.components.schemas import GPUTypeChoices
 from grandchallenge.core.guardian import get_objects_for_group
 from grandchallenge.core.models import RequestBase, UUIDModel
 from grandchallenge.core.storage import (
