+++ app/tests/evaluation_tests/test_models.py
@@ -11,7 +11,7 @@ from django.utils.timezone import now
 
 from grandchallenge.algorithms.models import Job
 from grandchallenge.components.models import ComponentInterface
-from grandchallenge.components.utils import GPUTypeChoices
+from grandchallenge.components.schemas import GPUTypeChoices
 from grandchallenge.evaluation.models import (
     SUBMISSION_WINDOW_PARENT_VALIDATION_TEXT,
     CombinedLeaderboard,
